+++ node18.nix
@@ -1,8 +1,0 @@
-# Minimal Supported Nodejs Version Environment
-{ pkgs ? import <nixpkgs> {} }:
-  pkgs.mkShell {
-    nativeBuildInputs = with pkgs; [
-      nodejs-18_x
-      yarn
-    ];
-}