+++ tests/errors/index.test.js
@@ -12,7 +12,7 @@ const target = name => path.join(__dirname, 'examples', name);
 test('[CORE] Realm execution timeout', async () => {
   const src = `Promise.resolve().then(() => { while(1); });`;
   try {
-    Script.execute(src, { run: { timeout: 1_000 } }, Script.contextify({}, true));
+    Script.execute(src, { runOpts: { timeout: 1_000 } }, Script.contextify({}, true));
     assert.fail('Should fail by timeout');
     // eslint-disable-next-line no-empty
   } catch (err) {}